requirements.txt
Django==1.11.2
django-bootstrap-form==3.2.1
django-enumfields==0.9.0
pbr==3.1.1
psycopg2==2.7.1
pytz==2017.2
six==1.10.0
stevedore==1.23.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
